% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnout.R
\docType{package}
\name{burnout}
\alias{burnout}
\title{burnout: A package for computing infectious disease persistence probabilities}
\description{
The burnout package uses analytical approximations to compute the
probabilities of fizzle and burnout in the stochastic SIR epidemic
model.
}
\section{The primary functions are ...}{
NA
}

