% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.R
\name{compare_q}
\alias{compare_q}
\title{Compare exact vs approximate calculations of Kendall's \eqn{q}}
\usage{
compare_q(
  R0 = NULL,
  Rmin = 1.001,
  Rmax = 64,
  nR0 = 101,
  epsilon = 10^(-(4:1)),
  nepsilon = length(epsilon),
  show.progress = FALSE
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{Rmin, Rmax}{minimum and maximum values of \eqn{{\cal R}_0}}

\item{nR0, nepsilon}{number of values of \code{R0} and \code{epsilon}}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{show.progress}{logical: if \code{TRUE} then spew each
element of the data frame to \code{stdout} as it is computed}
}
\value{
data frame with \eqn{\code{nR0*nepsilon}} rows
}
\description{
Compare exact vs approximate calculations of Kendall's \eqn{q}
}
\details{
The default \code{R0} value of \code{NULL} yields a
sensible sprinkling of \code{nR0} \eqn{{\cal R}_0} values.
}
\examples{
(cq <- compare_q(nR0=3))

}
\seealso{
\code{\link{q_exact}}, \code{\link{q_approx}}
}
