% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanH.R
\name{P1_prob_vanH}
\alias{P1_prob_vanH}
\title{Persistence probability via van Herwaarden (1997) approximation}
\usage{
P1_prob_vanH(R0, epsilon, k = 1, N = 10^6, subdivisions = 1000L, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{k}{initial number of infected individuals (\eqn{k})}

\item{N}{population size}

\item{subdivisions}{the maximum number of subintervals.}

\item{...}{additional arguments pass to \code{\link[stats]{integrate}}}
}
\description{
Persistence probability via van Herwaarden (1997) approximation
}
\seealso{
\code{\link{burnout_prob_vanH}}, \code{\link{P1_prob}}
}
