% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_prev.R
\name{peak_prev}
\alias{peak_prev}
\title{Peak prevalence}
\usage{
peak_prev(R0, epsilon = 0)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}
}
\value{
real number between 0 and 1
}
\description{
Approximation of peak prevalence in the SIR model with vital dynamics.
}
\details{
The peak prevalence is

\deqn{
  y_{\rm max} \approx 1-
    \dfrac{1}{{\cal R}_{0}}\big(1 +\ln{{\cal R}_{0}}\big)
    - \varepsilon 
  \Big(1-\dfrac{1}{{\cal R}_0}\Big)
  \dfrac{(\ln{{\cal R}_0})\big/\big(1-\frac{1}{{\cal R}_0}\big) - 1}
  {(\ln{{\cal R}_0})\big/\big(1-\frac{1}{{\cal R}_0}\big) + {\cal R}_0} \,.
}

The expression is exact for \eqn{\varepsilon = 0}.
}
\examples{
peak_prev(R0=2)
peak_prev(R0=2, epsilon=c(0, 0.001, 0.01, 0.1, 0.9))
peak_prev(R0=c(2,4), epsilon=0.1)
peak_prev(R0=c(2,4), epsilon=c(0,0.1)) # FIX: fails when both are vectors

}
