% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_prev.R
\name{peak_prev_nvd}
\alias{peak_prev_nvd}
\title{Peak prevalence}
\usage{
peak_prev_nvd(R0, epsilon = 0)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}
}
\value{
real number between 0 and 1
}
\description{
Exact peak prevalence in the SIR model with \emph{no vital dynamics} (nvd).
}
\details{
\deqn{
  y_{\rm max} = 1 -
    \dfrac{1}{{\cal R}_{0}}\big(1 +\ln{{\cal R}_{0}}\big) \,.
}

The optional argument (\code{epsilon}) is ignored if given.  The
reason for including it is so this function can be used as an
alternative to the more accurate \code{\link{peak_prev}} function
that does depend on \eqn{\varepsilon}.
}
\seealso{
\code{\link{peak_prev}}, \code{\link{x_in}}
}
