% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.compare_q}
\alias{plot.compare_q}
\title{Plot exact vs approximate \eqn{q}}
\usage{
\method{plot}{compare_q}(x, ...)
}
\arguments{
\item{x}{\code{\link{compare_q}} object, i.e., a
\code{\link{data.frame}} with particular structure}

\item{...}{additional parameters passed to \code{\link{plot}}}
}
\description{
\code{\link{plot}} method for \code{\link{compare_q}} objects.
}
\seealso{
\code{\link{compare_q}}, \code{\link{q_exact}},
\code{\link{q_approx}}
}
