% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_P1}
\alias{plot_P1}
\title{Plot persistence probability}
\usage{
plot_P1(epsilon = 0.01, N = 10^6, npts = 1001, ...)
}
\arguments{
\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{...}{additional parameters passed to \code{\link{plot}}}
}
\description{
Plot persistence probability
}
\examples{
plot_P1(epsilon=0.05, N=10^6)

}
