% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_herd.R
\name{plot_t_herd}
\alias{plot_t_herd}
\title{Plot herd immunity duration}
\usage{
plot_t_herd(
  R0 = exp(seq(log(1.001), log(2), length = 1001)),
  epsilon = c(0.01, 0.02, 0.03),
  col = 1:length(epsilon),
  lwd = 2,
  ylim = c(0, 60),
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{col}{vector of colours}
}
\description{
Plot herd immunity duration
}
\examples{
plot_t_herd(delta = 0.01)

}
\seealso{
\code{\link{t_herd}}
}
