% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{plot_x_in}
\alias{plot_x_in}
\title{Plot \eqn{x_{\rm in}}}
\usage{
plot_x_in(
  Rmin = 1.0001,
  Rmax = 20,
  R0 = exp(seq(log(Rmin), log(Rmax), length = 1001)),
  epsilon = c(0.01, 0.02, 0.03),
  xin_fun = x_in,
  col = 1:length(epsilon),
  lwd = 2,
  log = "x",
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{col}{vector of colours}
}
\description{
Plot \eqn{x_{\rm in}}
}
\examples{
plot_x_in(delta = 0.01)

}
\seealso{
\code{\link{x_in}}, \code{\link{x_in_exact}}
}
