% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.R
\name{q_approx}
\alias{q_approx}
\title{Kendall's \eqn{q} (Laplace approximation)}
\usage{
q_approx(R0, epsilon, xin = x_in(R0, epsilon))
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}
}
\value{
real number between 0 and 1
}
\description{
Approximate probability of eventual extinction in a birth-death process
starting from one individual.
}
\details{
Exploiting \href{https://en.wikipedia.org/wiki/Laplace\%27s_method}{Laplace's method} we
obtain

\deqn{
q({\cal R}_0,\varepsilon,x_{\rm in}) \approx
\dfrac{1}{1+\sqrt{\frac{2\pi}{\varepsilon ({\cal R}_{0}-1)}} \;e^{\frac{{\cal R}_{0}}{\varepsilon}\big(\frac{1}{{\cal R}_{0}}-x_{\rm in}\big)}
\Big(\frac{1-\frac{1}{{\cal R}_{0}}}{1-x_{\rm in}}\Big)^{\frac{{\cal R}_{0}}{\varepsilon}\big(1-\frac{1}{{\cal R}_{0}}\big)}}\,.
}
}
\examples{
q_approx(R0=2, epsilon=0.01)
q_approx(R0=2, epsilon=0)
q_approx(R0=c(2,4), epsilon=0.01)
q_approx(R0=2, epsilon=c(0.01, 0.1))
curve(q_approx(x,epsilon=0.01), from=1.01, to=2, las=1)

}
\references{
\insertRef{Kendall1948b}{burnout}
}
\seealso{
\code{\link{x_in}}, \code{\link{llig}}, \code{\link{q_exact}}
}
