% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_herd.R
\name{t_herd}
\alias{t_herd}
\title{Duration of herd immunity}
\usage{
t_herd(R0, epsilon, xin = x_in(R0, epsilon))
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{xin}{by default set via the approximation coded in \code{\link{x_in}}}

\item{delta}{probability threshold}
}
\description{
The time \eqn{t_{\rm H}} (after the first epidemic) from when the
proportion susceptible drops below the herd immunity threshold
(defined via \eqn{X(t_{\rm H}) = 1/{\cal R}_0}) until the
proportion susceptible next rises above the threshold.
}
\details{
\deqn{
	t_{\rm H} =  \dfrac{1}{\varepsilon}
    \ln\Big( \dfrac{1-x_{\rm in}}{1 - \frac{1}{{\cal R}_0}} \Big)
}
}
