% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanH.R
\name{vanH_prob}
\alias{vanH_prob}
\title{van Herwaarden (1997) approximation}
\usage{
vanH_prob(R0, epsilon, N = 10^6, subdivisions = 1000L)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{N}{population size}

\item{subdivisions}{the maximum number of subintervals.}
}
\description{
van Herwaarden (1997) approximation
}
\details{
See equation (8) of Ballard et al (2016)
}
