% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in_exact}
\alias{x_in_exact}
\title{Susceptibles at boundary layer (exact) (vector version)}
\usage{
x_in_exact(R0, epsilon, I0 = 1e-06)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{I0}{initial prevalence (proportion)}
}
\description{
Susceptibles at boundary layer (exact) (vector version)
}
