% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in_hocb}
\alias{x_in_hocb}
\title{Susceptibles at boundary layer (higher order corner/boundary layer approximation)}
\usage{
x_in_hocb(R0, epsilon, peakprev_fun = peak_prev, maxiter = 100, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{peakprev_fun}{function of \eqn{{\cal R}_0} and
\eqn{\varepsilon} to use to compute peak prevalence}

\item{maxiter}{maximum numbers of iterations for convergence of
\code{\link[emdbook]{lambertW}}}

\item{...}{additional arguments are ignored}
}
\description{
Susceptibles at boundary layer (higher order corner/boundary layer approximation)
}
\details{
\deqn{
	x_{\rm in} = 1+\left(1-\frac{1}{{\mathcal R}_{0}}\right)
  W_{-1}\left(-\dfrac{1-x_{\rm f}}{1-\frac{1}{{\mathcal R}_{0}}}\left(\dfrac{1-x_{\rm f}}{y^*}\dfrac{\frac{1}{{\mathcal R}_{0}}-x_{\rm f}}{x_{\rm f}}\right)^{\frac{\varepsilon}{{\mathcal R}_{0}}\frac{1}{1-\frac{1}{{\mathcal R}_{0}}}}
e^{-\frac{1-x_{\rm f}}{1-\frac{1}{{\mathcal R}_{0}}}-\varepsilon\frac{x_{\rm f}}{(1-x_{\rm f})\left(1-\frac{1}{{\mathcal R}_{0}}\right)}\tilde{Y}_{1}(x_{\rm f})}\right)
}
}
\seealso{
\code{\link{x_in}}, \code{\link{x_in_cb}}, \code{\link{x_in_exact}}
}
