% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{Ytilde_1_scalar}
\alias{Ytilde_1_scalar}
\title{\eqn{{\tilde Y}_1(x_{\rm f})}}
\usage{
Ytilde_1_scalar(xf, R0, tiny = 1e-12, subdivisions = 1000L, ...)
}
\arguments{
\item{tiny}{amount by which to avoid integration limits}
}
\description{
\eqn{{\tilde Y}_1(x_{\rm f})}
}
\details{
\deqn{
\tilde{Y}_{1}(x{\rm f}) =  \int_{x_{\rm f}}^{1} \left(\frac{1}{{\mathcal R}_{0}t} -1\right) \frac{1-t}{{\mathcal R}_{0} t Y_{0}(t)} + \frac{1-x_{\rm f}}{{\mathcal R}_{0}x_{\rm f}}\frac{1}{t-x_{\rm f}} \, dt
}
}
