% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnout.R
\docType{package}
\name{burnout}
\alias{burnout}
\title{burnout: A package for computing infectious disease persistence probabilities}
\description{
The burnout package uses analytical approximations to compute the
probabilities of fizzle and burnout in the stochastic SIR epidemic
model.
}
\section{Key functions}{


The primary functions are \code{\link{P1_prob}},
\code{\link{fizzle_prob}}, \code{\link{fizzle_time}}, and
\code{\link{t_herd}}.  There are also functions to compare
approximations of \eqn{{\mathscr P}_1} and component
calculations.
}

