% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.R
\name{compare_q}
\alias{compare_q}
\title{Compare exact and approximate calculations of Kendall's \eqn{q}}
\usage{
compare_q(
  n = 10,
  R0 = seq(1.01, 4, length = n),
  epsilon = seq(1e-04, 0.01, length = n),
  show.progress = TRUE
)
}
\arguments{
\item{n}{number of values of \code{R0} and \code{epsilon}}

\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{show.progress}{logical: if \code{TRUE} then spew each
element of the data frame to \code{stdout} as it is computed}
}
\value{
data frame with \eqn{\code{n}^2} rows
}
\description{
Compare exact and approximate calculations of Kendall's \eqn{q}
}
\examples{
(cq <- compare_q(n=3))
}
\seealso{
\code{\link{q_exact}}, \code{\link{q_approx}}
}
