% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.R
\name{compare_q}
\alias{compare_q}
\title{Compare exact and approximate calculations of Kendall's \eqn{q}}
\usage{
compare_q(
  n = 10,
  R0 = seq(1.01, 4, length = n),
  epsilon = seq(1e-04, 0.01, length = n),
  show.progress = TRUE
)
}
\description{
Compare exact and approximate calculations of Kendall's \eqn{q}
}
\examples{
(cq <- compare_q(n=3))
}
\seealso{
q_exact, q_approx
}
