% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_in.R
\name{x_in}
\alias{x_in}
\title{Susceptibles at boundary layer (approximation)}
\usage{
x_in(R0, epsilon, peakprev_fun = peak_prev, maxiter = 100, ...)
}
\arguments{
\item{R0}{basic reproduction number (\eqn{{\cal R}_0})}

\item{epsilon}{mean infectious period as a proportion of mean
lifetime (\eqn{\varepsilon = 0} corresponds to infinite
lifetime, i.e., no mortality)}

\item{peakprev_fun}{function of \eqn{{\cal R}_0} and
\eqn{\varepsilon} to use to compute peak prevalence}

\item{maxiter}{maximum numbers of iterations for convergence of
\code{\link[emdbook]{lambertW}}}

\item{...}{additional arguments are ignored}
}
\value{
real number between 0 and 1
}
\description{
Fraction of hosts that are susceptible when the trajectory enters
the boundary layer at the end of the first epidemic.
}
\details{
\deqn{
	x_{\rm in} = -\dfrac{1}{{\cal R}_{0}} W_{0}\left(-{\cal R}_{0} e^{-{\cal R}_{0}\left(1-y^{\star}\right)}\right) + \varepsilon\, e^{{\cal R}_{0} y^{\star}}\big(E_{1}({\cal R}_{0} y^{\star}) - E_{1}({\cal R}_{0} y_{\rm max}) \big) \,.
}

Here, \eqn{W_{0}} denotes the principal branch of the Lambert
\eqn{W} function
(\insertCite{Olver2010;textual}{burnout}, \insertCite{Corless1996;textual}{burnout},
\href{https://dlmf.nist.gov/4.13}{NIST 4.13}),
\eqn{E_{1}(x) =
\int_{x}^{\infty} \frac{e^{-t}}{t}\,{\rm d}t} is the exponential
integral function (\insertCite{Olver2010;textual}{burnout}, \href{https://dlmf.nist.gov/6.2}{NIST 6.2}),
\eqn{y^\star=\varepsilon(1-\frac{1}{{\cal R}_0})}
is the equilibrium prevalence,
and \eqn{y_{\rm max}} is the \link[=peak_prev]{peak prevalence}.
}
\examples{
x_in(R0=2, epsilon=0.01)
x_in(R0=2, epsilon=0)
x_in(R0=c(2,4), epsilon=0.01)
x_in(R0=2, epsilon=c(0.01, 0.1))
curve(x_in(x,epsilon=0.001), from=1.01, to=5, las=1, n=1001)
curve(x_in(x,epsilon=0.01), from=1.01, to=5, las=1, add=TRUE, col="magenta", n=1001)
curve(x_in(x,epsilon=0.1), from=1.01, to=5, las=1, add=TRUE, col="cyan", n=1001)

}
\references{
\insertRef{Corless1996}{burnout}

\insertRef{Olver2010}{burnout}
}
\seealso{
\code{\link{peak_prev}}, \code{\link[emdbook]{lambertW}},
\code{\link[expint]{expint_E1}}
}
